backend/requirements.txt
Django==6.0.3
djangorestframework==3.17.1
psycopg2-binary==2.9.11
environs==15.0.0
gunicorn==25.3.0
djangorestframework-simplejwt==5.5.1